requirements.txt
# pip list
# pip show

# pip install -r /path/to/requirements.txt

# https://realpython.com/twitter-bot-python-tweepy/#deploying-bots-to-a-server-using-docker

# Posting
tweepy==4.12.1
mastodon.py==1.8.0

# General python
pandas==1.5.3
numpy==1.24.2
scipy==1.10.1
python-dotenv==0.21.1

# Plotting
matplotlib==3.7.0
plotly==5.13.0

# Scraping
sodapy==2.2.0
beautifulsoup4==4.11.2
lxml==4.9.2

# Weather - in progress
zarr==2.14.0
s3fs
xarray

# Cloud Storage
google-cloud-storage==2.7.0
# pydrive2==1.14.0
